+++ src/alien_invasion.py
@@ -6,11 +6,10 @@ import pygame
 from alien import Alien
 from bullet import Bullet
 from game_stats import GameStats
-# from laser import Laser
 from settings import Setting
 from ship import Ship
 from bonus import Bonus
-# from guarder import Gaurder
+from bonus_system import BonusSystem
 
 
 class AlienInvasion:
@@ -32,13 +31,12 @@ class AlienInvasion:
         self.stats = GameStats(self)
 
         self.ship = Ship(self)
-        self.bullets = pygame.sprite.Group()
-        # self.laser = Laser(self)
-        # self.guarder = Gaurder(self)
         self.bonus = Bonus(self)
-        self.aliens = pygame.sprite.Group()
+        self.bonus_system = BonusSystem(self)
+        self.alien = Alien(self)
+        self.bullet = Bullet(self)
         # create fleet
-        self._create_fleet()
+        self.alien._create_fleet()
         # Set the background color.
         self.bg_color = (230, 230, 230)
 
@@ -49,12 +47,9 @@ class AlienInvasion:
 
             if self.stats.game_active:
                 self.ship.update()
-                self._update_bullets()
-                self._update_aliens()
-                self._update_bonus()
-                # self.laser._update_laser(self)
-                # self.guarder._update_guarder(self)
-                self.bonus._update(self)
+                self.bullet._update_bullets(self)
+                self.alien._update_aliens(self)
+                self.bonus_system._update(self)
             self._update_screen()
 
     def _check_events(self):
@@ -68,13 +63,11 @@ class AlienInvasion:
     def _check_keydown_events(self, event):
         """Respond to keypress. """
         if event.key == pygame.K_l:
-            # self.laser.fire_flag = True
-            self.bonus.laser.fire_flag = True
+            self.bonus_system.laser.fire_flag = True
         if event.key == pygame.K_g:
-            # self.guarder.fire_flag = True
-            self.bonus.guarder.fire_flag = True
+            self.bonus_system.guarder.fire_flag = True
         if event.key == pygame.K_SPACE:
-            self._fire_bullet()
+            self.bullet._fire_bullet()
         if event.key == pygame.K_RIGHT:
             self.ship.moving_right = True
         if event.key == pygame.K_LEFT:
@@ -89,95 +82,9 @@ class AlienInvasion:
         if event.key == pygame.K_LEFT:
             self.ship.moving_left = False
         if event.key == pygame.K_l:
-            # self.laser.fire_flag = False
-            self.bonus.laser.fire_flag = False
+            self.bonus_system.laser.fire_flag = False
         if event.key == pygame.K_g:
-            # self.guarder.fire_flag = False
-            self.bonus.guarder.fire_flag = False
-
-    def _fire_bullet(self):
-        """create a new bullet and add it to the bullets group."""
-        if len(self.bullets) < self.settings.bullets_allowed:
-            new_bullet = Bullet(self)
-            self.bullets.add(new_bullet)
-
-    def _update_bullets(self):
-        """update position of bullets and get rid of old bullets."""
-
-        self.bullets.update()
-        # get rid of bullets that have disappeared.
-        for bullet in self.bullets.copy():
-            if bullet.rect.bottom <= 0:
-                self.bullets.remove(bullet)
-        self._check_bullet_alien_collisions()
-
-    def _check_bullet_alien_collisions(self):
-        """respond to bullet-alien collisions."""
-        # remove any bullets and aliens that have collided.
-        conllisions = pygame.sprite.groupcollide(
-            self.bullets, self.aliens, True, True)
-        if not self.aliens:
-            # destroy exiting bullets and create new fleet.
-            self.bullets.empty()
-            self._create_fleet()
-
-    def _create_fleet(self):
-        """create the fleet of aliens."""
-        # create an alien and find the number of aliens in a row.
-        # spacing between each alien is equal to one alien width.
-        alien = Alien(self)
-        alien_width, alien_height = alien.rect.size
-
-        available_space_x = self.settings.screen_width - 2 * alien_width
-        number_aliens_x = available_space_x // (2 * alien_width)
-
-        # determine the number of rows of aliens that fit on the screen.
-        ship_height = self.ship.rect.height
-        available_space_y = self.settings.screen_height - \
-            ship_height - (3 * alien_height)
-        number_rows = available_space_y // (2 * alien_height)
-
-        # create the full fleet of aliens.
-        for row_number in range(number_rows):
-            for alien_number in range(number_aliens_x):
-                self._create_alien(alien_number, row_number)
-
-    def _create_alien(self, alien_number, row_number):
-        """create an alien and place it in the row."""
-        alien = Alien(self)
-        alien_width, alien_height = alien.rect.size
-        alien.x = alien_width + 2 * alien_width * alien_number
-        alien.rect.x = alien.x
-        alien.rect.y = alien_height + 2 * alien.rect.height * row_number
-        self.aliens.add(alien)
-
-    def aliens_shoot_bullets(self):
-        """respond aliens shoot bullets to ship"""
-        pass
-
-    def _check_fleet_edges(self):
-        """respond appropriately if any aliens have reached an edge."""
-        for alien in self.aliens.sprites():
-            if alien.check_edges():
-                self._change_fleet_direction()
-                break
-
-    def _change_fleet_direction(self):
-        """drop the entire fleet and change the fleet's direction."""
-        for alien in self.aliens.sprites():
-            alien.rect.y += self.settings.fleet_drop_speed
-        self.settings.fleet_direction *= -1
-
-    def _update_aliens(self):
-        """Check if the fleet is at an edge, then update the positions of all aliens in the fleet."""
-        self._check_fleet_edges()
-        self.aliens.update()
-        # look for alien-ship collisions.
-        if pygame.sprite.spritecollideany(self.ship, self.aliens):
-            self._ship_hit()
-
-        # aliens hitting the bottom of the screen
-        self._check_aliens_bottom()
+            self.bonus_system.guarder.fire_flag = False
 
     def _ship_hit(self):
         """Respond to the ship being hit by an alien."""
@@ -185,12 +92,12 @@ class AlienInvasion:
             # decrement ships_left.
             self.stats.ship_left -= 1
             # Get rid of any remaining aliens and bullets.
-            self.aliens.empty()
-            self.bullets.empty()
+            self.alien.aliens.empty()
+            self.bullet.bullets.empty()
             self.bonus.bonuses.empty()
 
             # Create a new fleet and center the ship.
-            self._create_fleet()
+            self.alien._create_fleet()
             self.ship.center_ship()
 
             # pause.
@@ -198,33 +105,12 @@ class AlienInvasion:
         else:
             self.stats.game_active = False
 
-    def _check_aliens_bottom(self):
-        """Check if any aliens have reached the bottom of the screen."""
-        screen_rect = self.screen.get_rect()
-        for alien in self.aliens.sprites():
-            if alien.rect.bottom >= screen_rect.bottom:
-                # Treat this the same as if the ship got hit.
-                self._ship_hit()
-                break
-
-    def _update_bonus(self):
-        """update position of bonuses and get rid of old bonuses."""
-        self.bonus.check_edges()
-        self.bonus.update(self)
-        # look for collision between ship and bonus
-        self.bonus._check_bonus_ship_collisions(self)
-        # bonuses hitting the bottom of the screen
-        self.bonus._check_bonus_bottom()
-
     def _update_screen(self):
         """Update images on the screen, and flip to the new ship."""
         self.screen.fill(self.settings.bg_color)
         self.ship.blitme()
-        for bullet in self.bullets.sprites():
+        for bullet in self.bullet.bullets.sprites():
             bullet.draw_bullet()
-        # self.laser.draw()
-        self.aliens.draw(self.screen)
-        self.bonus.draw_bonus()
-        self.bonus._draw()
-        # self.guarder.draw()
+        self.alien.aliens.draw(self.screen)
+        self.bonus_system._draw()
         pygame.display.flip()